# Optimize JSON serialization