* optimize ObjectInstance.GetOwnPropertyKeys for non-sorting case
* increase StringBuilderPool max capacity check
* smarter checks for integer numbers in JsonParser

## Jint.Benchmark.JsonBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Parse|bestbuy_dataset.json|297.742 ms|5.7835 ms|115.01 MB|
| **New** |	|	| **299.811 ms (+1%)** | **4.3964 ms** | **114.15 MB (-1%)** |
| Old |Stringify|bestbuy_dataset.json|142.319 ms|2.2880 ms|104.63 MB|
| **New** |	|	| **127.802 ms (-10%)** | **1.9511 ms** | **90.18 MB (-14%)** |
| Old |Parse|twitter.json|3.023 ms|0.0057 ms|3.64 MB|
| **New** |	|	| **2.995 ms (-1%)** | **0.0111 ms** | **3.61 MB (-1%)** |
| Old |Stringify|twitter.json|2.926 ms|0.0241 ms|2.74 MB|
| **New** |	|	| **2.434 ms (-17%)** | **0.0118 ms** | **1.4 MB (-49%)** |


/cc @tomatosalat0